+++ src/components/battle/voteResult/types.ts
@@ -1,0 +1,6 @@
+export interface Vote {
+  title: string;
+  albumCoverUrl: string;
+  selectedPostVoteCnt: number;
+  oppositePostVoteCnt: number;
+}
